+++ apps/meteor/app/livechat/server/lib/RoutingManager.ts
@@ -10,6 +10,7 @@ import type {
 	InquiryWithAgentInfo,
 	TransferData,
 } from '@rocket.chat/core-typings';
+import { License } from '@rocket.chat/license';
 import { Logger } from '@rocket.chat/logger';
 import { LivechatInquiry, LivechatRooms, Subscriptions, Rooms, Users } from '@rocket.chat/models';
 import { Match, check } from 'meteor/check';
@@ -80,6 +81,13 @@ export const RoutingManager: Routing = {
 			this.methodName = 'Manual_Selection';
 		} else {
 			this.methodName = name;
+		}
+
+		const shouldPreventQueueStart = await License.shouldPreventAction('monthlyActiveContacts');
+
+		if (shouldPreventQueueStart) {
+			logger.error('Monthly Active Contacts limit reached. Queue will not start');
+			return;
 		}
 
 		void (await Omnichannel.getQueueWorker()).shouldStart();
@@ -178,10 +186,6 @@ export const RoutingManager: Routing = {
 			return false;
 		}
 
-		if (!(await Omnichannel.isWithinMACLimit(room))) {
-			throw new Error('error-mac-limit-reached');
-		}
-
 		if (departmentId && departmentId !== department) {
 			logger.debug(`Switching department for inquiry ${inquiry._id} [Current: ${department} | Next: ${departmentId}]`);
 			await updateChatDepartment({
@@ -269,10 +273,6 @@ export const RoutingManager: Routing = {
 	},
 
 	async transferRoom(room, guest, transferData) {
-		if (!(await Omnichannel.isWithinMACLimit(room))) {
-			throw new Error('error-mac-limit-reached');
-		}
-
 		logger.debug(`Transfering room ${room._id} by ${transferData.transferredBy._id}`);
 		if (transferData.departmentId) {
 			logger.debug(`Transfering room ${room._id} to department ${transferData.departmentId}`);