+++ apps/meteor/client/views/admin/subscription/components/cards/ActiveSessionsPeakCard.tsx
@@ -7,6 +7,7 @@ import { useFormatDate } from '../../../../../hooks/useFormatDate';
 import { useStatistics } from '../../../../hooks/useStatistics';
 import type { CardProps } from '../FeatureUsageCard';
 import FeatureUsageCard from '../FeatureUsageCard';
+import UpgradeButton from '../UpgradeButton';
 
 const ActiveSessionsPeakCard = (): ReactElement => {
 	const { t } = useTranslation();
@@ -24,13 +25,17 @@ const ActiveSessionsPeakCard = (): ReactElement => {
 		title: t('ActiveSessionsPeak'),
 		infoText: t('ActiveSessionsPeak_InfoText'),
 		...(exceedLimit && {
-			upgradeButtonText: t('Upgrade'),
+			upgradeButton: (
+				<UpgradeButton target='active-session-peak-card' action='upgrade' small>
+					{t('Upgrade')}
+				</UpgradeButton>
+			),
 		}),
 	};
 
 	return (
 		<FeatureUsageCard card={card}>
-			{!isLoading && maxMonthlyPeakConnections ? (
+			{!isLoading && maxMonthlyPeakConnections !== undefined ? (
 				<Box textAlign='center'>
 					<Box fontScale='h1' color={exceedLimit ? 'font-danger' : 'font-default'}>
 						{used} / {total}